+++ splash/SplashMath.h
@@ -172,7 +172,7 @@ static inline int splashRound(SplashCoord x)
     __asm fldcw WORD PTR oldCW;
     return result;
 #else
-    return (int)splashFloor(x + 0.5);
+    return splashFloor(x + 0.5);
 #endif
 }
